pyproject.toml
[tool.poetry]
name = "lesson-24-1"
version = "0.1.0"
description = ""
authors = ["max_slidzen <slidzon94@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_24"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
